Section Updates Fixes + Adult Content
git-svn-id: http://aptoide.org/repo/aptoide-client@1761 8d55eb69-6847-4b75-98a7-935b3913b0ec

+++ unstable/Aptoidev5/trunk/v5engine/src/main/java/cm/aptoide/ptdev/downloadmanager/DownloadInfo.java
@@ -5,6 +5,7 @@ import android.os.Environment;
 import android.preference.PreferenceManager;
 import android.util.Log;
 
+import android.widget.ListView;
 import cm.aptoide.ptdev.Aptoide;
 import cm.aptoide.ptdev.downloadmanager.event.DownloadEvent;
 import cm.aptoide.ptdev.downloadmanager.event.DownloadStatusEvent;
@@ -258,12 +259,15 @@ public class DownloadInfo implements Runnable, Serializable {
                     fileToDelete = file;
                 }
 
+
             }
             if (fileToDelete != null) {
                 Log.d("TAG", "Deleting " + fileToDelete.getName());
-                fileToDelete.delete();
+                if(!fileToDelete.delete()){
+                    return;
+                }
+                checkDirectorySize(dirPath);
             }
-            checkDirectorySize(dirPath);
         }
     }
 
@@ -326,6 +330,7 @@ public class DownloadInfo implements Runnable, Serializable {
     public void remove() {
         changeStatusState(new CompletedState(this));
 
+        if (mFilesToDownload == null) return;
         for (DownloadModel model : mFilesToDownload) {
             new File(model.getDestination()).delete();
         }
